refactored, added helper file for awardAmountChart
